chore: sever client 분리
